requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.47.0
django-environ==0.8.0
django-on-heroku==1.1.2
gunicorn==20.1.0
idna==3.3
json2html==1.3.0
oauthlib==3.1.1
Pillow==8.4.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycparser==2.21
PyJWT==2.3.0
python-decouple==3.5
python-openid==2.2.5
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.1
urllib3==1.26.7
whitenoise==5.3.0
